// src/lib/services/models/DailyBudgetTipViewChitStandardResponse.ts
/* generated using openapi-typescript-codegen -- do not edit */
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */
import type { DailyBudgetTipView } from './DailyBudgetTipView';
import type { HttpStatusCode } from './HttpStatusCode';
export type DailyBudgetTipViewChitStandardResponse = {
    href?: string | null;
    relations?: Array<string> | null;
    method?: string | null;
    routeName?: string | null;
    routeValues?: any;
    status?: boolean;
    message?: string | null;
    data?: DailyBudgetTipView;
    statusCode?: HttpStatusCode;
    errors?: any;
};